% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNA.TEvMP.R
\name{microRNA.TEvMP}
\alias{microRNA.TEvMP}
\title{A function to run the microRNA.TEvMP shiny app}
\usage{
microRNA.TEvMP(...)
}
\arguments{
\item{...}{Any argument that you can pass to shiny::runApp}
}
\value{
Plot

Runs the microRNA.TEvMP shiny app.
}
\description{
This function runs the microRNA.TEvMP app
}
\examples{
# microRNA.TEvMP()
}
