% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRplot.R
\name{miRplot}
\alias{miRplot}
\title{Get path to readData example}
\usage{
miRplot(df, color.groups, shape.groups)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
Located example excel file in package
}
\description{
readData comes bundled with a example files in its `inst/app/www`
directory. This function makes them easy to access.
}
\examples{

}
